# 95 schedule # 94 password

+++ src/main/java/defencer/controller/UserProfileController.java
@@ -6,23 +6,28 @@ import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXDatePicker;
 import com.jfoenix.controls.JFXTextField;
 import defencer.data.CurrentUser;
-import defencer.model.DaysOff;
-import defencer.model.Instructor;
-import defencer.model.Project;
-import defencer.model.Schedule;
+import defencer.model.*;
 import defencer.service.InstructorService;
 import defencer.service.factory.ServiceFactory;
 import defencer.util.NotificationUtil;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Label;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
+import javafx.stage.Window;
+import lombok.SneakyThrows;
 
 import java.net.URL;
 import java.sql.SQLException;
@@ -39,6 +44,8 @@ public class UserProfileController implements Initializable {
     private JFXButton btnSure;
     @FXML
     private JFXButton btnCanNot;
+    @FXML
+    private JFXButton btnChangePassword;
     @FXML
     private TableView<Schedule> tableMyProjects;
     @FXML
@@ -64,8 +71,6 @@ public class UserProfileController implements Initializable {
     @FXML
     private JFXTextField email;
     @FXML
-    private Label status;
-    @FXML
     private JFXButton btnSetDaysBusy;
     @FXML
     private JFXButton btnDeleteDayOff;
@@ -78,7 +83,6 @@ public class UserProfileController implements Initializable {
 
     private ObservableList<DaysOff> observableDaysOff = FXCollections.observableArrayList();
     private ObservableList<Schedule> observableMyProject = FXCollections.observableArrayList();
-
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -101,6 +105,7 @@ public class UserProfileController implements Initializable {
         btnUpdateProfile.setOnAction(s -> updateProfile());
         btnCanNot.setOnAction(e -> disagree());
         btnSure.setOnAction(e -> agree());
+        btnChangePassword.setOnAction(this::changePassword);
     }
 
     /**
@@ -205,7 +210,10 @@ public class UserProfileController implements Initializable {
      * Set status free for instructor if he cat't take the project.
      */
     private void disagree() {
-        Schedule schedule = tableMyProjects.getSelectionModel().getSelectedItem();
+        final Schedule schedule = tableMyProjects.getSelectionModel().getSelectedItem();
+        if (schedule == null) {
+            return;
+        }
         ServiceFactory.getWiseacreService().deleteSelectedInstructors(CurrentUser.getLink().getId(), schedule.getProjectId());
         CurrentUser.refresh(CurrentUser.getLink().getEmail());
         loadTableMyProject();
@@ -215,17 +223,12 @@ public class UserProfileController implements Initializable {
      * Set status busy if instructor take the project.
      */
     private void agree() {
-        final CurrentUser currentUser = CurrentUser.getLink();
-        if (currentUser.getProjectId() == null) {
+        final Schedule schedule = tableMyProjects.getSelectionModel().getSelectedItem();
+        if (schedule == null) {
             return;
         }
-        try {
-            ServiceFactory.getWiseacreService().updateCurrentUser(currentUser.getId(), "BUSY");
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        CurrentUser.refresh(CurrentUser.getLink().getEmail());
-        factoryInitialization();
+        ServiceFactory.getWiseacreService().updateSchedule(CurrentUser.getLink(), schedule);
+        loadTableMyProject();
     }
 
     /**
@@ -233,20 +236,27 @@ public class UserProfileController implements Initializable {
      */
     private void factoryInitialization() {
         CurrentUser currentUser = CurrentUser.getLink();
-
         firstName.setText(currentUser.getFirstName());
         lastName.setText(currentUser.getLastName());
         phone.setText(currentUser.getPhoneNumber());
         email.setText(currentUser.getEmail());
-        status.setText(currentUser.getStatus());
-
-        initializeProject(currentUser);
     }
 
     /**
-     * @param currentUser for initialize project.
+     * Change password for current user.
      */
-    private void initializeProject(CurrentUser currentUser) {
-
+    @SneakyThrows
+    private void changePassword(ActionEvent event) {
+        final FXMLLoader fxmlLoader = new FXMLLoader();
+        fxmlLoader.setLocation(getClass().getResource("/NewPass.fxml"));
+        Parent parent = fxmlLoader.load();
+        final Stage stage = new Stage();
+        Scene value = new Scene(parent);
+        stage.setScene(value);
+        value.getStylesheets().add("css/main.css");
+        stage.initModality(Modality.WINDOW_MODAL);
+        Window window = ((Node) event.getSource()).getScene().getWindow();
+        stage.initOwner(window);
+        stage.show();
     }
 }
